requirements.txt
airtest==1.2.10
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cached-property==1.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
decorator==5.1.1
Deprecated==1.2.13
docker==4.4.4
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
facebook-wda==1.4.6
ffmpeg-python==0.2.0
filelock==3.12.0
Flask==2.3.2
future==0.18.3
gitdb2==2.0.6
GitPython==2.1.15
hrpc==1.0.9
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.2
mss==6.1.0
nose==1.3.7
numpy==1.24.3
opencv-contrib-python==4.6.0.66
packaging==21.3
paramiko==3.1.0
Pillow==9.5.0
poco==0.99.1
pocoui==1.0.89
py==1.11.0
pyaml==19.12.0
pycparser==2.21
PyGithub==1.45
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-gitlab==1.15.0
python-xlib==0.33
pywinauto==0.6.3
PyYAML==5.3
requests==2.30.0
retry==0.9.2
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
svn==1.0.1
texttable==1.6.7
urllib3==2.0.2
websocket-client==0.48.0
Werkzeug==2.3.4
wrapt==1.15.0
